Display only what's relevant
This backports a patch in the 22.04 release that filters out a variety
of useless sandbox-related mounts to clean up the display. In addition,
it limits what the app can see to only the user's home directory, as
the display of the host's root filesystem is still exposed in a rather
messy way to Flatpak, and arguably this is more appropriate for a
sandboxed app context anyway.